Implement rough code editor
