+++ .vscode/settings.json
@@ -15,5 +15,6 @@
     "editor.formatOnSave": false
   },
   "tslint.autoFixOnSave": true,
-  "typescript.tsdk": "./node_modules/typescript/lib"
+  "typescript.tsdk": "./node_modules/typescript/lib",
+  "debug.node.autoAttach": "on"
 }
