+++ www/gatsby-browser.js
@@ -1,10 +1,10 @@
-// @ts-check
-
 /**
  * Implement Gatsby's Browser APIs in this file.
  *
  * See: https://www.gatsbyjs.org/docs/browser-apis/
  */
+
+// @ts-check
 
 import { Boot } from "./src/Boot"
