+++ www/gatsby-config.js
@@ -1,5 +1,7 @@
 // @ts-check
 
+const playgroundHastPlugin = require("./src/utils/playgroundHastPlugin")
+
 module.exports = {
   siteMetadata: {
     title: "Palette",
@@ -7,6 +9,7 @@ module.exports = {
     author: "Artsy",
   },
   plugins: [
+    "gatsby-plugin-catch-links",
     {
       resolve: "gatsby-source-filesystem",
       options: {
@@ -19,10 +22,14 @@ module.exports = {
     {
       resolve: "gatsby-mdx",
       options: {
+        extensions: [".mdx", ".md"],
         defaultLayouts: {
           posts: require.resolve("./src/components/Layout.tsx"),
           default: require.resolve("./src/components/Layout.tsx"),
         },
+        hastPlugins: [playgroundHastPlugin],
+        // mdPlugins: [],
+        // gatsbyRemarkPlugins: [{}],
       },
     },
   ],
